chore (PS-51): restructure test folders into unit and integration
