+++ point-system/src/test/java/pointsystem/integration/PointSystemApplicationTests.java
@@ -1,4 +1,4 @@
-package pointsystem;
+package pointsystem.integration;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
