+++ point-system/src/test/java/pointsystem/unit/controller/EmployeeControllerTest.java
@@ -1,4 +1,4 @@
-package pointsystem.controller.EmployeeControllerTests;
+package pointsystem.unit.controller;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
